remove special webauth scenario
